refactor(scripts): Swift Regex in the facade generator, lint-clean layout
- NSRegularExpression + NSRange bridging replaced with typed regex
  literals (#/.../#) — drops both force-tries and the index conversions.
- URL(fileURLWithPath:) → URL(filePath:).
- Oneof emission moved to an Emitter extension around a small OneofContext,
  and the string/bytes field cases extracted — the emitter and its longest
  functions now fit swiftlint's type/function length limits on merit;
  only file_length is disabled (swift-sh scripts are single-file by design).
- Nanopb.swift split: field accessors (strings, bytes, scalars, enums,
  submessage views, oneof releases, repeated) now live in
  NanopbAccessors.swift; the core runtime (NanopbBox, NanopbMessage,
  wire-format primitives) stays put.

Generated output is unchanged (same 135 emitted types, byte-identical
apart from the CocoaPods import guard).
